marge M8.sh

+++ M8.sh
@@ -13,14 +13,28 @@ cd $(dirname $0)
 #echo "powersave" | sudo tee /sys/devices/system/cpu/cpu*/cpufreq/scaling_governor
 #echo "performance" | sudo tee /sys/devices/system/cpu/cpu*/cpufreq/scaling_governor
 
-# alsaloop_wait will run in background and try to create the loopback 
-# if it can't be created before m8c runs. Useful for wait_for_device=true. 
-#./alsaloop_wait &
-alsaloop -P hw:rockchiprk817co,0 -C hw:M8,0 -t 200000 -A 5 --rate 44100 --sync=1 -T -1 -d
-sleep 1
-./_m8c/m8c
+aplay -l | grep 'rockhiprk817co'
+ret = $?
 
-pkill alsaloop
+if [ ret -eq 0]
+then
+#ArkOS
+    alsaloop -P hw:rockchiprk817co,0 -C hw:M8,0 -t 200000 -A 5 --rate 44100 --sync=1 -T -1 -d
+    sleep 1
+    ./_m8c/m8c
+    pkill alsaloop
+else
+#ROCKNIX
+    M8=$(pa-info 2>&1 | awk '/Name: alsa_input\.usb-DirtyWave_M8_[0-9\-].*\.analog-stereo/{print $2;exit}')
+    SPK=$(pa-info 2>&1 | awk '/Name: alsa_output\._sys_devices_platform_.*HiFi__Headphones__sink/{print $2;exit}')
+    pactl set-default-source ${M8}
+    pactl set-default-sink ${SPK}
+    pactl load-module module-loopback source=${M8} sink=${SPK} latency_msec=200 format=s16le rate=44100 channels=2
+    sleep 2
+    ./_m8c/m8c
+    pactl unload-module module-loopback
+fi
+
 
 # Uncomment the following lines to enable wi-fi
 #sudo modprobe -i mt7601u
